Split grammar files following Boost Spirit conventions
This will allow rules to be shared in the future (ex: interpolating
variables on double quoted strings or strings as parameters to commands)
.

+++ command/Command.cpp
@@ -16,7 +16,7 @@ CommandStore& CommandStore::instance() {
 namespace {
   string_view make_view(string::const_iterator begin,
                         string::const_iterator end){
-    return {&*begin, distance(begin, end)};
+    return {&*begin, static_cast<size_t>(distance(begin, end))};
   }
 }
